Grafana (#996)
* Add grafana to install


* Adjust influx start order

Start influx prior to postgres check to ensure migration can happen in future.
Ensure grafana DB is created in postgres
Add localhost as trusted for chroma connections to postgres


* Setup grafana CONF_FILE

Set CONF_FILE in /etc/sysconfig/grafana-server


* fix rpm build


* Fixups


* Reload systemctl before enabling grafana


* Use postgresql socket

+++ chroma_core/lib/service_config.py
@@ -42,6 +42,7 @@ from chroma_core.lib.util import CommandLine, CommandError
 from iml_common.lib.ntp import NTPConfig
 from iml_common.lib.firewall_control import FirewallControl
 from iml_common.lib.service_control import ServiceControl, ServiceControlEL7
+from iml_common.lib.util import wait_for_result
 
 log = logging.getLogger("installation")
 try:
@@ -286,7 +287,7 @@ class ServiceConfig(CommandLine):
         # Enable use of the management plugin if its available, else this tag is just ignored.
         self.try_shell(sudo + ["rabbitmqctl", "set_user_tags", settings.AMQP_BROKER_USER, "management"])
 
-    def _setup_influxdb_service(self):
+    def _setup_influxdb(self):
         influx_service = ServiceControlEL7("influxdb")
 
         log.info("Starting InfluxDB...")
@@ -303,9 +304,37 @@ class ServiceConfig(CommandLine):
             log.error(error)
             raise RuntimeError(error)
 
-    def _setup_influxdb_db(self):
+        # Wait for influx to finish starting
+        wait_for_result(
+            lambda: self.try_shell(["influx", "-execute", "exit"]),
+            logger=log,
+            timeout=60,
+            expected_exception_classes=[CommandError],
+        )
+
         log.info("Creating InfluxDB databse...")
         self.try_shell(["influx", "-execute", "CREATE DATABASE iml"])
+
+    def _setup_grafana(self):
+        cfg_file = "/etc/sysconfig/grafana-server"
+        if os.path.exists("%s.dist" % cfg_file):
+            return
+        shutil.copy2(cfg_file, "%s.dist" % cfg_file)
+        with open(cfg_file, "a") as fn:
+            fn.write("CONF_FILE=/etc/grafana/grafana-iml.ini")
+
+        # grafana needs daemon-reload before enable and start
+        self.try_shell(["systemctl", "daemon-reload"])
+        service = ServiceControlEL7("grafana-server")
+        error = service.enable()
+        if error:
+            log.error(error)
+            raise RuntimeError(error)
+        error = service._start()
+        if error:
+            log.error(error)
+            raise RuntimeError(error)
+        return
 
     def _setup_crypto(self):
         if not os.path.exists(settings.CRYPTO_FOLDER):
@@ -393,7 +422,8 @@ class ServiceConfig(CommandLine):
     @staticmethod
     def _config_pgsql_auth(database):
         auth_cfg_file = "/var/lib/pgsql/data/pg_hba.conf"
-        os.rename(auth_cfg_file, "%s.dist" % auth_cfg_file)
+        if not os.path.exists("%s.dist" % auth_cfg_file):
+            os.rename(auth_cfg_file, "%s.dist" % auth_cfg_file)
         with open(auth_cfg_file, "w") as cfg:
             # Allow our django user to connect with no password
             cfg.write("local\tall\t%s\t\ttrust\n" % database["USER"])
@@ -401,6 +431,12 @@ class ServiceConfig(CommandLine):
             cfg.write("local\tall\tall\t\tident\n")
 
     PathStats = namedtuple("PathStats", ["total", "used", "free"])
+
+    def _try_psql_sql(self, sql):
+        return self.try_shell(["su", "postgres", "-c", 'psql -tAc "%s"' % sql])
+
+    def _psql_sql(self, sql):
+        return self.shell(["su", "postgres", "-c", 'psql -tAc "%s"' % sql])
 
     def _path_space(self, path):
         """Returns the disk statistics of the given path.
@@ -429,17 +465,20 @@ class ServiceConfig(CommandLine):
             log.error(error_msg)
             return error_msg
 
+    def _restart_pgsql(self):
+        postgresql_service = ServiceControl.create("postgresql")
+        postgresql_service.restart()
+        postgresql_service.enable()
+
     def _setup_pgsql(self, database, check_db_space):
         log.info("Setting up PostgreSQL service...")
 
         self._init_pgsql(database)
 
-        postgresql_service = ServiceControl.create("postgresql")
-        postgresql_service.restart()
-        postgresql_service.enable()
+        self._restart_pgsql()
 
         tries = 0
-        while self.shell(["su", "postgres", "-c", "psql -c '\\d'"])[0] != 0:
+        while self._psql_sql("\\d")[0] != 0:
             if tries >= 4:
                 raise RuntimeError("Timed out waiting for PostgreSQL service to start")
             tries += 1
@@ -454,20 +493,12 @@ class ServiceConfig(CommandLine):
             log.info("Creating database owner '%s'...\n" % database["USER"])
 
             # Enumerate existing roles
-            _, roles_str, _ = self.try_shell(["su", "postgres", "-c", "psql -t -c 'select " "rolname from pg_roles;'"])
+            _, roles_str, _ = self._try_psql_sql("select rolname from pg_roles")
             roles = [line.strip() for line in roles_str.split("\n") if line.strip()]
 
             # Create database['USER'] role if not found
             if not database["USER"] in roles:
-                self.try_shell(
-                    [
-                        "su",
-                        "postgres",
-                        "-c",
-                        "psql -c 'CREATE ROLE %s NOSUPERUSER "
-                        "CREATEDB NOCREATEROLE INHERIT LOGIN;'" % database["USER"],
-                    ]
-                )
+                self._try_psql_sql("CREATE ROLE %s NOSUPERUSER CREATEDB NOCREATEROLE INHERIT LOGIN;" % database["USER"])
 
             log.info("Creating database '%s'...\n" % database["NAME"])
             self.try_shell(["su", "postgres", "-c", "createdb -O %s %s;" % (database["USER"], database["NAME"])])
@@ -554,17 +585,22 @@ class ServiceConfig(CommandLine):
 
     def _setup_database(self, check_db_space):
         error = None
-        if not self._db_accessible():
-            # For the moment use the builtin configuration
-            # TODO: this is where we would establish DB name and credentials
-            databases = settings.DATABASES
+        # For the moment use the builtin configuration
+        # TODO: this is where we would establish DB name and credentials
+        databases = settings.DATABASES
 
+        if not self._db_accessible():
             error = self._setup_pgsql(databases["default"], check_db_space)
+
         else:
-            log.info("DB already accessible")
+            log.info("Postgres already accessible")
 
         if error:
             return error
+
+        _, out, _ = self._try_psql_sql("SELECT datname FROM pg_catalog.pg_database WHERE datname = 'grafana'")
+        if "grafana" not in out:
+            self.try_shell(["su", "postgres", "-c", "createdb -O %s grafana;" % (databases["default"]["USER"])])
 
         self._syncdb()
 
@@ -733,6 +769,8 @@ proxy=_none_
 
         self.set_nginx_config()
 
+        self._setup_influxdb()
+
         error = self._setup_database(check_db_space)
         if check_db_space and error:
             return [error]
@@ -745,12 +783,11 @@ proxy=_none_
         self._setup_ntp(ntp_server)
         self._setup_crypto()
 
-        self._setup_influxdb_service()
         self._setup_rabbitmq_service()
 
         self._setup_rabbitmq_credentials()
 
-        self._setup_influxdb_db()
+        self._setup_grafana()
 
         self._enable_services()
         self._start_services()
@@ -792,7 +829,9 @@ proxy=_none_
 
         # Check services are active
         interesting_services = (
-            self.MANAGER_SERVICES + self.CONTROLLED_SERVICES + ["postgresql", "rabbitmq-server", "influxdb"]
+            self.MANAGER_SERVICES
+            + self.CONTROLLED_SERVICES
+            + ["postgresql", "rabbitmq-server", "influxdb", "grafana-server"]
         )
 
         service_config = self._service_config(interesting_services)
